+++ src/backtesting.rs
@@ -12,7 +12,7 @@ use crate::risk::{calculate_risk, RiskCalculationErrors};
 use crate::strategies::Strategy;
 use crate::types::{Candle, ExecutedTrade, FutureTrade, MarketData, MarketDataError, Side, Signal};
 use crate::utils;
-use crate::utils::{AlignmentError, check_candle_alignment, print_candle_statistics, trim_candles};
+use crate::utils::{AlignmentError, check_candle_alignment, extract_candles_from_df, print_candle_statistics, trim_candles};
 
 const CANDLE_TRIM_SIZE: IdxSize = 100;
 
@@ -119,7 +119,7 @@ impl BacktestingRuntime {
         // compute indicator graph
         let trading_candles = self.get_trading_asset()?.to_owned();
 
-        self.strategy.process_candles(&trading_candles).unwrap();
+        self.strategy.process_candle(&trading_candles).unwrap();
 
         // populate market and trading candles
         self.trading_candles = trading_candles.into();
@@ -165,21 +165,23 @@ impl BacktestingRuntime {
         // initialize position manager
         let mut position_manager = PositionManager::new(self.manager_config.clone());
 
-        // process historical data
-        let (relevant_candles, signals) = self.get_relevant_signals()?;
+        let candle_rows = extract_candles_from_df(self.trading_candles.as_ref().unwrap()).unwrap();
 
         // begin trading simulation
         let start_time = Instant::now();
-        for (candle, signal) in relevant_candles.iter().zip(signals.iter()) {
+        for candle in candle_rows {
             let trimmed_trading_candles = trim_candles(self.trading_candles.as_ref().unwrap(), candle.time, CANDLE_TRIM_SIZE);
             if trimmed_trading_candles.height() == 0 {
                 continue;
             }
-            let trimmed_candles = utils::extract_candles_from_df(&trimmed_trading_candles).unwrap();
+            let signal = self.strategy.process_candle(&trimmed_trading_candles)
+                .map_err(|_| BacktestingErrors::SignalExtractionError)?;
+
+            let trimmed_candles = extract_candles_from_df(&trimmed_trading_candles).unwrap();
 
             // trim market data
             let trimmed_market = trim_candles(&self.market_candles.as_ref().unwrap(), candle.time, CANDLE_TRIM_SIZE);
-            let trimmed_market = utils::extract_candles_from_df(&trimmed_market).unwrap();
+            let trimmed_market = extract_candles_from_df(&trimmed_market).unwrap();
 
             // calculate current portfolio risk metrics
             let risk = calculate_risk(&portfolio, &trimmed_market, &trimmed_candles)
@@ -191,7 +193,7 @@ impl BacktestingRuntime {
             let current_price = candle.close;
 
             // make decision based on risk, signals and current market conditions
-            let decision = position_manager.make_decision(&mut portfolio, &risk, signal, current_price)
+            let decision = position_manager.make_decision(&mut portfolio, &risk, &signal, current_price)
                 .map_err(|e| {
                     info!("Error making decision: {:?}", e);
                     BacktestingErrors::DecisionError(e)
@@ -218,62 +220,6 @@ impl BacktestingRuntime {
         self.print_statistics(elapsed, &portfolio);
 
         Ok(())
-    }
-
-    /// Write all indicator graphs to a specific dir
-    ///
-    /// # Arguments
-    /// * `dir` - The path to save the indicator graph files
-    pub fn save_indicator_data(&mut self, dir: &str) -> Result<(), PolarsError> {
-        let dir_path = Path::new(dir);
-
-        for indicator in self.strategy.indicators.iter_mut() {
-            let indicator = indicator.as_mut();
-            let file_name = format!("{}_graph.csv", indicator.get_name());
-            let path = dir_path.join(file_name);
-            let path = path.to_str().unwrap();
-            indicator.save_graph_as_csv(path)?;
-        }
-        Ok(())
-    }
-
-    /// This is used to run the backtesting simulation on rows which are relevant.
-    ///
-    /// Generate relevant rows signals and return values from which to iterate over
-    ///
-    /// Signals are generated for the current candles, the candles are joined with the signals.
-    /// A combined iterator of signals which are not `Hold` and candles is returned.
-    ///
-    ///
-    /// # Arguments
-    /// * `candles` - Candles to use for signal generation
-    fn get_relevant_signals(&mut self) -> Result<(Vec<Candle>, Vec<Signal>), BacktestingErrors> {
-        // generate signals
-        let signals =  self.strategy.get_combined_signals()
-            .map_err(|_| BacktestingErrors::SignalExtractionError)?;
-        let signals = signals.unwrap();
-
-        // join signals column with candles df
-        let combined = self.trading_candles
-            .clone()
-            .unwrap()
-            .lazy()
-            .join(
-                signals.lazy(),
-                [col("time")],
-                [col("time")],
-                JoinArgs::new(JoinType::Left),
-            )
-            .collect().unwrap();
-
-        // ensure that columns are correct
-        assert_eq!(combined.get_column_names(), &["time", "open", "high", "low", "close", "volume", "signals"]);
-
-        // convert to vector of candles and vector of signals
-        let candles_vec = utils::extract_candles_from_df(&combined).unwrap();
-        let sides: Vec<Signal> = utils::extract_signals_from_df(&combined, "signals").unwrap();
-
-        Ok((candles_vec, sides))
     }
 
     /// Create a portfolio from the [`PortfolioArgs`]
@@ -317,9 +263,6 @@ impl BacktestingRuntime {
         // save market data
         let path = format!("data/{}_{}.csv", self.trading_config.market_asset, self.trading_config.frequency);
         save_candles(self.market_candles.as_mut().unwrap(), &path).unwrap();
-
-        // save strategy indicator data
-        self.save_indicator_data("data").unwrap()
     }
 }